Refactor unit tests and add a few more test cases
